+++ src/Serwisant/SerwisantApi/Types/SchemaService/CustomFieldValue.php
@@ -8,12 +8,20 @@ use Serwisant\SerwisantApi\Types;
 class CustomFieldValue extends Types\Type
 {
   /**
+   * @var string
+   * Primary key of particular field value to be used for update. NULL value means it's a field created after this entity creation.
+  */
+  public $ID;
+
+  /**
    * @var CustomField
+   * Entity of field definition.
   */
   public $field;
 
   /**
    * @var string
+   * Value of field for this particular object - it's always a string here - look into `field` definition to get proper casting.
   */
   public $value;