+++ src/Serwisant/SerwisantApi/Types/SchemaService/Customer.php
@@ -23,6 +23,12 @@ class Customer extends Types\Type
   public $addresses;
 
   /**
+   * @var CustomerAgreementValue[]
+   * Customer agreements, including any new agreement customer should take (this is complete agreements list)
+  */
+  public $agreements;
+
+  /**
    * @var File
   */
   public $avatar;